pyproject.toml
[project]
name = "cbse-question-paper-generator"
version = "0.1.0"
description = "Deep Agent for generating CBSE question papers from JSON blueprints"
requires-python = ">=3.11,<3.14"
readme = "README.md"
authors = [
    {name = "CBSE Question Paper Generator", email = "info@example.com"}
]
license = {text = "MIT"}
keywords = ["cbse", "question-paper", "deep-agents", "langchain", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Education",
]

dependencies = [
    "deepagents>=0.3.12",
    "langchain>=1.2.7,<1.3.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    # Qdrant vector database for question retrieval
    "qdrant-client>=1.12.0",
    # Fuzzy matching for topic matching
    "rapidfuzz>=3.0.0",
    # Settings management
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
    "rich>=14.3.1",
    "typer>=0.15.0",
    # Diagram and DOCX generation
    "drawsvg>=2.4.1",
    "cairosvg>=2.7.0",
    "python-docx>=1.1.0",
    # Async and cache
    "aiosqlite>=0.20.0",
    "aiofiles>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "black>=25.0.0",
    "ruff>=0.9.0",
    "mypy>=1.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "src/**/*.py",
    "src/**/*.md",
    "src/**/*.json",
    "AGENTS.md",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py311", "py312", "py313"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
strict_optional = true

[project.scripts]
cache-cli = "src.cache.cache_cli:app"
